añadido modal con angular material para confirmar el borrado de notas

+++ src/app/components/note/note.component.ts
@@ -1,8 +1,11 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
 
 import { NotesService } from '../../services/notes.service';
 import { NoteModel } from '../../models/note';
+
+import { ConfirmDeletionComponent } from '../confirm-deletion/confirm-deletion.component';
 
 @Component({
   selector: 'note',
@@ -14,18 +17,25 @@ export class NoteComponent {
   @Input() noteModel!: NoteModel;
   @Output() noteDeleted: EventEmitter<void> = new EventEmitter();
 
-  constructor(private notesService: NotesService) {}
+  constructor(private notesService: NotesService, private dialog: MatDialog ) {}
 
   deleteNote() {
-    if(this.noteModel._id) { // Comprobación necesaria, pues sin ella hay un error ya que la id podría ser undefined al llamar a la función de eliminado
-        this.notesService.deleteNote(this.noteModel._id).subscribe({
-          next: (deletedNote) => {
-            this.noteDeleted.emit(); // Para notificar al componente notes-list y poder actualizar la lista de notas
-          },
-          error: (err) => console.error('Error al eliminar la nota:', err)
-        });
-    } else {
-      console.error("No se ha encontrado una nota con ese ID")
-    }
+    // Se ejecuta el componente confirm-deletion como modal
+    const dialogRef = this.dialog.open(ConfirmDeletionComponent);
+
+    // Se comprueba si en el modal se ha confirmado el eliminado y entonces se procede con él
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        if (this.noteModel._id) {
+          this.notesService.deleteNote(this.noteModel._id).subscribe({
+            next: () => {
+              this.noteDeleted.emit();
+            },
+            error: (err) => console.error('Error al eliminar la nota:', err),
+          });
+        }
+      }
+    });
   }
+
 }
